Fix QT_SUPPORTED_WP_VERSION
After wordpress core was updated, we needed to change
QT_SUPPORTED_WP_VERSION to current version of WP.
In this version of qTranslate I'm using global function $wp_version. So,
it will work with the new versions of Wordpress, and there will be no
need to change version every time.